update self monitoring and metric recorders

+++ inspectit-ocelot-config/src/main/java/rocks/inspectit/ocelot/config/model/metrics/definition/MetricDefinitionSettings.java
@@ -42,7 +42,7 @@ public class MetricDefinitionSettings {
 
     @NotNull
     @Builder.Default
-    private InstrumentValueType valueType = InstrumentValueType.DOUBLE;
+    private InstrumentValueType valueType = InstrumentValueType.LONG;
 
     /**
      * The description of the metric.
